backend/sky-agent-py/pyproject.toml
[project]
name = "sky-agent-py"
version = "0.1.0"
description = "Merchant operations analysis agent for sky-take-out"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "httpx>=0.28.0,<1.0.0",
    "pydantic-settings>=2.6.0,<3.0.0",
    "uvicorn>=0.32.0,<1.0.0",
]

[dependency-groups]
dev = []
